Add another workaround for rspec-puppet testing
As a follow up to d6c401045103f299ff3cf9cbdeb2ea2abe554f98 we need to
bypass actual decryption when using onceover and `node_encrypt.secret`.

This is needed because rspec-puppet resolves Deferred functions.
See https://github.com/rodjek/rspec-puppet/pull/743

We also can't just mock this function in onceover, because the normal
way to mock a ruby function is with a puppet language function, but
these do not work with `Deferred`.