(CONT-1048) - Skip broken node_encrypt::file test
Skipping this test due to a difference in behaviour between mocha stub
and rspec allow when calling the PuppetX::Binford2k::NodeEncrypt.decrypt
method on Ruby 2.x only.

Voted by the team to skip, as test case will be removed in
https://github.com/puppetlabs/puppetlabs-node_encrypt/pull/94 with the
removal of the deprecated node_encrypt::file defined type.